refactor(sync-merge-proptest): split helpers into sync_merge_proptest_helpers/ submodule
The all-four-properties file landed at 780 LOC — well over the project's
500-LOC soft cap (per `feedback_split_files_proactively`). Extract the
fixture constants, record builders (`live_record`, `tombstoned_record`),
fixture builders (`build_no_veto_fixture`, `build_single_veto_fixture`,
`build_two_veto_fixture`), the `VetoDecision` constructor helper
(`make_decision`), the unlocked-identity opener (`open_identity`), the
`sync_once → ConcurrentDetected` driver, and the canonical-block reader
into `core/tests/sync_merge_proptest_helpers/mod.rs`.

Result: `sync_merge_proptest.rs` drops to 391 LOC (four `proptest!`
blocks + their per-property docstrings), helpers module is 370 LOC. Both
comfortably under the cap; the proptest file's intent is now solely the
property bodies + their rationale.

No behavioural change. Gauntlet on the refactor: workspace cargo
795 / 0 / 10 (unchanged), clippy clean, fmt clean, conformance PASS,
spec-citation freshness PASS.
